requirements.txt
streamlit
supabase
pdfplumber
openai
python-dotenv
